Rename autoconfig package to autodiscovery
