+++ feature/account/setup/src/main/kotlin/app/k9mail/feature/account/setup/ui/autodiscovery/item/ConfigurationApprovalItem.kt
@@ -1,10 +1,10 @@
-package app.k9mail.feature.account.setup.ui.autoconfig.item
+package app.k9mail.feature.account.setup.ui.autodiscovery.item
 
 import androidx.compose.foundation.lazy.LazyItemScope
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import app.k9mail.feature.account.setup.domain.input.BooleanInputField
-import app.k9mail.feature.account.setup.ui.autoconfig.view.ConfigurationApprovalView
+import app.k9mail.feature.account.setup.ui.autodiscovery.view.ConfigurationApprovalView
 import app.k9mail.feature.account.setup.ui.common.item.ListItem
 
 @Composable
