+++ feature/account/setup/src/main/kotlin/app/k9mail/feature/account/setup/ui/autodiscovery/item/EmailAddressItem.kt
@@ -1,4 +1,4 @@
-package app.k9mail.feature.account.setup.ui.autoconfig.item
+package app.k9mail.feature.account.setup.ui.autodiscovery.item
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.lazy.LazyItemScope
@@ -6,26 +6,28 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import app.k9mail.core.common.domain.usecase.validation.ValidationError
-import app.k9mail.core.ui.compose.designsystem.molecule.input.PasswordInput
+import app.k9mail.core.ui.compose.designsystem.molecule.input.EmailAddressInput
 import app.k9mail.feature.account.setup.ui.common.item.ListItem
 import app.k9mail.feature.account.setup.ui.common.toResourceString
 
 @Composable
-internal fun LazyItemScope.PasswordItem(
-    password: String,
+internal fun LazyItemScope.EmailAddressItem(
+    emailAddress: String,
     error: ValidationError?,
-    onPasswordChange: (String) -> Unit,
+    onEmailAddressChange: (String) -> Unit,
     modifier: Modifier = Modifier,
+    isEnabled: Boolean = true,
 ) {
     val resources = LocalContext.current.resources
 
     ListItem(
         modifier = modifier,
     ) {
-        PasswordInput(
-            password = password,
+        EmailAddressInput(
+            emailAddress = emailAddress,
             errorMessage = error?.toResourceString(resources),
-            onPasswordChange = onPasswordChange,
+            onEmailAddressChange = onEmailAddressChange,
+            isEnabled = isEnabled,
             contentPadding = PaddingValues(),
         )
     }
