+++ feature/account/setup/src/main/kotlin/app/k9mail/feature/account/setup/ui/autodiscovery/item/PasswordItem.kt
@@ -1,4 +1,4 @@
-package app.k9mail.feature.account.setup.ui.autoconfig.item
+package app.k9mail.feature.account.setup.ui.autodiscovery.item
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.lazy.LazyItemScope
@@ -6,28 +6,26 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import app.k9mail.core.common.domain.usecase.validation.ValidationError
-import app.k9mail.core.ui.compose.designsystem.molecule.input.EmailAddressInput
+import app.k9mail.core.ui.compose.designsystem.molecule.input.PasswordInput
 import app.k9mail.feature.account.setup.ui.common.item.ListItem
 import app.k9mail.feature.account.setup.ui.common.toResourceString
 
 @Composable
-internal fun LazyItemScope.EmailAddressItem(
-    emailAddress: String,
+internal fun LazyItemScope.PasswordItem(
+    password: String,
     error: ValidationError?,
-    onEmailAddressChange: (String) -> Unit,
+    onPasswordChange: (String) -> Unit,
     modifier: Modifier = Modifier,
-    isEnabled: Boolean = true,
 ) {
     val resources = LocalContext.current.resources
 
     ListItem(
         modifier = modifier,
     ) {
-        EmailAddressInput(
-            emailAddress = emailAddress,
+        PasswordInput(
+            password = password,
             errorMessage = error?.toResourceString(resources),
-            onEmailAddressChange = onEmailAddressChange,
-            isEnabled = isEnabled,
+            onPasswordChange = onPasswordChange,
             contentPadding = PaddingValues(),
         )
     }
